moved add picture function from ContactItem to SidebarNavBar
