+++ .gitignore
@@ -9,7 +9,7 @@ node_modules/
 
 # misc
 .DS_Store
-.env
+client/.env
 .env.local
 .env.development.local
 .env.test.local
