--- client/.env
@@ -1,3 +1,0 @@
-REACT_APP_ACCESS_KEY_ID="AKIAXH26RTZDQQILKIXQ"
-REACT_APP_SECRET_KEY_ID="nAu9BnXaXZXFMkYHmibGBh682k+fBhWsZ8fDKvSt"
-REACT_APP_BUCKET_NAME="peppermint-1234
