+++ client/src/components/ContactItem/index.js
@@ -9,21 +9,18 @@ import {
   Menu,
   MenuItem,
 } from "@material-ui/core";
-import { DropzoneDialog } from "material-ui-dropzone";
 import { MoreHoriz } from "@material-ui/icons";
-import uploadUserImage from "../../services/uploadUserImage";
 
-const ContactItem = ({ imageUrl, name, index, isOnline }) => {
-  const [open, setOpen] = useState(false);
+const ContactItem = ({
+  imageUrl,
+  name,
+  index,
+  isOnline,
+  handleDeleteContactButton,
+}) => { 
   const [anchorEl, setAnchorEl] = useState(null);
 
   const classes = useStyles();
-
-  const handleSave = (files) => {
-    uploadUserImage(files[0]);
-    setAnchorEl(null);
-    setOpen(false);
-  };
 
   const handleClick = (event) => {
     setAnchorEl(event.currentTarget);
@@ -61,22 +58,10 @@ const ContactItem = ({ imageUrl, name, index, isOnline }) => {
         open={Boolean(anchorEl)}
         onClose={handleClose}
       >
-        <MenuItem onClick={() => setOpen(true)}>Add picture</MenuItem>
-        <MenuItem onClick={() => console.log("Contact deleted")}>
+        <MenuItem onClick={() => handleDeleteContactButton("ya@ya.ru", index)}>
           Delete contact
         </MenuItem>
-      </Menu>
-
-      <DropzoneDialog
-        open={open}
-        onSave={handleSave}
-        showPreviews={true}
-        maxFileSize={300000}
-        onClose={() => {
-          setOpen(false);
-          handleClose();
-        }}
-      />
+      </Menu>      
     </ListItem>
   );
 };
@@ -85,7 +70,7 @@ export default memo(ContactItem);
 
 ContactItem.propTypes = {
   name: PropTypes.string.isRequired,
-  imageUrl:PropTypes.string,
-  index:PropTypes.number.isRequired,
-  isOnline:PropTypes.bool.isRequired,
+  imageUrl: PropTypes.string,
+  index: PropTypes.number.isRequired,
+  isOnline: PropTypes.bool,
 };
