+++ client/src/containers/SidebarNavBar/index.js
@@ -1,8 +1,10 @@
 import React, { useState } from "react";
 import { Typography, ButtonBase, Menu, MenuItem } from "@material-ui/core";
 import { useStyles } from "./style";
+import { DropzoneDialog } from "material-ui-dropzone";
 import { MoreHoriz } from "@material-ui/icons";
 import UserAvatar from "../../components/UserAvatar/index";
+import uploadUserImage from "../../services/uploadUserImage";
 
 const SidebarNavBar = () => {
   const [open, setOpen] = useState(false);
@@ -16,6 +18,12 @@ const SidebarNavBar = () => {
 
   const handleClose = () => {
     setAnchorEl(null);
+  };
+
+  const handleSave = (files) => {
+    uploadUserImage(files[0]);
+    setAnchorEl(null);
+    setOpen(false);
   };
 
   return (
@@ -42,10 +50,19 @@ const SidebarNavBar = () => {
           open={Boolean(anchorEl)}
           onClose={handleClose}
         >
-          <MenuItem onClick={() => console.log("Logged out")}>
-            Log out
-          </MenuItem>
+          <MenuItem onClick={() => setOpen(true)}>Add picture</MenuItem>
+          <MenuItem onClick={() => console.log("Logged out")}>Log out</MenuItem>
         </Menu>
+        <DropzoneDialog
+          open={open}
+          onSave={handleSave}
+          showPreviews={true}
+          maxFileSize={300000}
+          onClose={() => {
+            setOpen(false);
+            handleClose();
+          }}
+        />
       </div>
     </div>
   );
