+++ client/src/context/contacts/contactsReducer.js
@@ -8,7 +8,9 @@ export default function contactsReducer(state, action) {
     case DELETE_CONTACT: {
       return {
         ...state,
-        contacts: state.contacts.splice(action.payload.index, 1),
+        contacts: state.contacts
+          .slice(0, action.payload.index)
+          .concat(state.contacts.slice(action.payload.index + 1)),
       };
     }
     default:
