+++ server/routes/contacts.js
@@ -3,7 +3,7 @@ const data = require("../data-modules/dataService")();
 router.get("/:email/contacts", async (req, res) => {
   try {
     //find contacts by user id
-    const user = await data.getUserByEmail(req.params.email);
+    const user = await data.getContacts(req.params.email);
 
     //response with contacts
     res.status(200).send(user);
